+++ spec/kitchen_spec/directions/bake_chapter_references/v2_spec.rb
@@ -18,7 +18,7 @@ RSpec.describe Kitchen::Directions::BakeChapterReferences::V2 do
         <div data-type="page" id="page1">
           <h1 data-type='document-title'>Introduction to Sociology</h1>
           <p>Some text
-            <a href="#something" data-type="cite" id="some-id">
+            <a href="#note1" data-type="cite" id="some-id">
               <div data-type="note" class="delete-me">
                 <!-- no-selfclose -->
               </div>
@@ -27,7 +27,7 @@ RSpec.describe Kitchen::Directions::BakeChapterReferences::V2 do
           <section data-depth="1" id="1" class="reference">
             <h3 data-type="title">Endnotes</h3>
             <div data-type="note" id="note1">
-              <a href="#note1">
+              <a href="#some-id">
                 <span>1</span>
               </a>
               <span>Victoria Knowles, “What’s the Difference Between CSR and Sustainability?”</span>
@@ -37,7 +37,7 @@ RSpec.describe Kitchen::Directions::BakeChapterReferences::V2 do
         <div data-type="page" id="page2">
           <h1 data-type='document-title'>What Is Sociology?</h1>
           <p>Some text
-            <a href="#something2" data-type="cite" id="some-id2">
+            <a href="#note2" data-type="cite" id="some-id2">
               <div data-type="note" class="delete-me">
                 <!-- no-selfclose -->
               </div>
@@ -46,7 +46,7 @@ RSpec.describe Kitchen::Directions::BakeChapterReferences::V2 do
           <section data-depth="1" id="2" class="reference">
             <h3 data-type="title">Endnotes</h3>
             <div data-type="note" id="note2">
-              <a href="#note2">
+              <a href="#some-id2">
                 <span>2</span>
               </a>
               <span>Abercrombie, Nicholas, Stephen Hill, and Bryan S. Turner. 2000. The Penguin Dictionary of Sociology. London: Penguin.</span>